consolidate meta tags to the agency data component (#294)
