+++ frontend/src/Components/Charts/TrafficStops/TrafficStops.js
@@ -38,7 +38,6 @@ import useDataset, { STOPS_BY_REASON, STOPS, AGENCY_DETAILS } from '../../../Hoo
 import { P, WEIGHTS } from '../../../styles/StyledComponents/Typography';
 
 // Hooks
-import useMetaTags from '../../../Hooks/useMetaTags';
 import useTableModal from '../../../Hooks/useTableModal';
 
 // Children
@@ -112,7 +111,6 @@ function TrafficStops(props) {
     ],
   });
 
-  const renderMetaTags = useMetaTags();
   const [renderTableModal, { openModal }] = useTableModal();
 
   const [stopPurposeGroupsData, setStopPurposeGroups] = useState({
@@ -709,7 +707,6 @@ function TrafficStops(props) {
   return (
     <TrafficStopsStyled>
       {/* Traffic Stops by Percentage */}
-      {renderMetaTags()}
       {renderTableModal()}
       <div style={{ display: 'flex', justifyContent: 'center' }}>
         <MonthRangePicker
